-Se agregaron services y controllers de: Administrador,Alumno,Carrera,Comision,Materia,Profesor con sus respectivos metodos para hacer CRUDs -Se agrego GlobalExceptionHandler y la excepcion EntidadNoEncontradaException -Se cambio a la estrategia de herencia JOINED tanto en la bdd como en java para la relacion Persona - Alumno,Profesor y Administrador

+++ src/main/java/com/sistema/sistema/entities/areaAcademica/Carrera.java
@@ -15,7 +15,7 @@ public class Carrera {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idCarrera;
+    private Long idCarrera;
 
     @Column(length = 50)
     private String nombre;
